+++ sciris/sc_version.py
@@ -4,6 +4,6 @@ Version and license information.
 
 __all__ = ['__version__', '__versiondate__', '__license__']
 
-__version__      = '3.2.3'
-__versiondate__  = '2025-07-05'
+__version__      = '3.2.4'
+__versiondate__  = '2025-07-12'
 __license__      = f'Sciris {__version__} ({__versiondate__}) – © 2014-2025 by the Sciris Development Team'